+++ src/lib/AsyncTypeaheadInput.tsx
@@ -27,6 +27,7 @@ const AsyncTypeaheadInput = <T extends FieldValues>(props: AsyncTypeaheadProps<T
     addonRight,
     className = "",
     style,
+    emptyLabel,
   } = props;
   const { name, id } = useSafeNameId(props.name, props.id);
 
@@ -88,6 +89,7 @@ const AsyncTypeaheadInput = <T extends FieldValues>(props: AsyncTypeaheadProps<T
             onFocus={focusHandler}
             {...reactBootstrapTypeaheadProps}
             style={style}
+            emptyLabel={emptyLabel}
           />
         </FormGroupLayout>
       )}