+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `emptyLabel` property to `TypeaheadInput` components
+
 ## [2.2.2] - 2023-12-01
 
 ### Added